requirements.txt
discord.py==1.3.3
dynaconf[yaml]==2.2.3
peewee==3.13.2
requests==2.23.0
beautifulsoup4==4.9.0
aiocron==1.3